build: don't always rebuild .vapi file
vapigen not touching nor overwrite the dest file if its content hasn't
changed. This causes the rule to generate the .vapi file to always trigger
as it will always be older than the .gir file which is generated
during the build. Workaround this issue by touching the .vapi file
after generating it.

https://bugzilla.gnome.org/show_bug.cgi?id=683257