GdStack: Change how we crossfade
We used to just draw the current child to an offscreen pixbuf
when we switched to a new page, but that is not really correct
as it often happens that the child got queue_resize() on it
so the draw fails due to alloc_needed.

So, we completely change how we do this. In the case of an
animated transition we keep both the old and the new child
child_visible, then we expose the old child *once* in the normal
draw handler and then keep that pixmap around for the rest
of the transition.