Introduce GdHeaderButton
GdHeaderButton is a simple interface for buttons on header bars and main
toolbars.
The interface allows to get/set a label, a symbolic icon name, and
control whether the label should use Pango markup.

Three base classes are available, for simple, toggle and menu buttons,
which will automatically set appropriate style classes according to the
property values and the type of button.