GdTaggedEntry: Fix inserting tag when unmapped
When a GdTaggedEntryTag was inserted into an unmapped but realized
GdTaggedEntry, the tag's GDK window wasn't created. This resulted in a
drawing failure on showing the entry again. Fix this by making sure to
realize the tag at once if the entry is already realized.

Also enhance test-tagged-entry to allow testing this.

https://bugzilla.gnome.org/show_bug.cgi?id=712755