GdTaggedEntryTag: Fix close button states
When clicking on a tag's close button, the close buttons of all tags
where drawn with the active state (well, except the button you
actually clicked on, which kept showing the hover state), because
of a wrong if-clause and setting two states at the same time.

I'm not sure if it is necessary to avoid setting GTK_STATE_FLAG_ACTIVE
and GTK_STATE_FLAG_PRELIGHT at the same time on the tag buttons, I
guess ACTIVE should take precedence here, but I'll leave that to the
CSS experts to figure out.

https://bugzilla.gnome.org/show_bug.cgi?id=720764