view: Make selection changes instant
If the view model was a filter model, the selection icon would only
appear or disappear on mouse hover, when the tree view was fetching the
value of the selected column again. This makes it look like selection
is horribly slow.

This patch makes sure that the icon view is instantly updated.