Don't include git.mk in makefile
We have a .gitignore committed into version control, so we shouldn't be
overwriting it. This defeats the purpose of using git.mk as we wind up
with a dirty subproject after every build due to make rewriting our
tracked .gitignore. And libgd itself does not use git.mk, so it
shouldn't be here.

https://bugzilla.gnome.org/show_bug.cgi?id=769616