main-view: Don't let the Cairo surface get bigger while copying
The cairo_surface_create_similar API expects the width and height to be
in device-space units, not pixels. So either we need to scale down the
outputs of cairo_image_surface_get_{width, height}, or we need to use
cairo_surface_create_similar_image. I prefer the latter because we are
already using other image surface API and pixel values.

Note that unlike cairo_surface_create_similar, we need to manually set
the device scale because cairo_surface_create_similar_image doesn't
inherit it.

https://bugzilla.gnome.org/show_bug.cgi?id=776133