meson: Move all generation and compilation to libgd/
In the process, remove headers files from the list of sources. This is
not necessary with Meson.

The move is required for compiling resources with
gnome.compile_resources() which was also missing earlier

Also, just use '.' to include_directories since that resolves to both
the builddir and the srcdir, which is exactly what we want since some of
the headers we use are generated and hence in the builddir.