main-icon-box-child: Remove the margin-bottom from the GtkOverlay
The margin-bottom was preventing the icon from being centred, and it
looks bad when the grid is letterboxed.

The intention was to add some spacing between the icon and the first
row of text. However, current users of this widget are not showing any
text, and this is best done by the theme.

https://bugzilla.gnome.org/show_bug.cgi?id=779032