main-box, main-icon-box: Emit selection-changed with updated selection
Sometimes, GdMainBox::selection-changed is emitted before the list of
selected child widgets has been updated. This can happen when a
selected GdMainBoxItem was removed from the GListModel.

GdMainBox was monitoring removals from the GListModel in order to emit
selection-changed. However, since it had connected to
GListModel::items-changed before GdMainIconBox, it was emitting the
signal before the list of selected children got updated. This can be
problematic for an application.

An application should be able to use gd_main_box_get_selection from
its GdMainBox::selection-changed handler and have the updated list of
selected child widgets.

One option is to have GdMainBox connect to the GListModel after
GdMainIconBox. However, it would be better to move this logic to
GdMainIconBox which deals with all the other cases where
selection-changed needs to be emitted. It is easier to ensure that the
implementation of a single class stays consistent, as opposed to
trying to keep two different classes in sync.

GtkFlowBox, the parent of GdMainIconBox, uses the GtkContainer::removed
virtual method to handle this case. Therefore, it is a natural choice
for us to hook in and emit the signal.

https://bugzilla.gnome.org/show_bug.cgi?id=779589