gd-two-lines-renderer: use Pango alpha attribute
This code was previously trying to alter the color of the second
line by manipulating the foreground color. Now that gtk+ 3.22
is our stable branch, we can rely on more recent stable features
such as Pango's alpha attribute.

The problem with the code previously is that it would not
respect the GTK_CELL_RENDERER_SELECTED state. It would draw the
altered non-selected state with a dim-level on top of a
selected row.

By simply avoiding the foreground color altogether (and
inheriting it from the PangoLayout state when rendering), we
get the appropriate color and also blend into the selected row
state properly.