libgd.m4: Fix order for gtk-hacks
Order gtk-hacks below widgets requiring it in the macro.

Apparently the order of conditional dependencies matter for m4 macros.
For example, main-view is depending on gtk-hacks, but gtk-hacks is
defined before main-view and did not get picked up. This resulted in
linking errors when gtk-hacks itself was omitted from the superprojects
init.