main-icon-view: Explicitly specify the vertical alignment of the text
The text rendered by the TwoLinesRenderer is vertically aligned to
touch its border with the TogglePixbufRenderer. This was only working
this way because TwoLinesRenderer was ignoring its y-offset and always
treating it as zero.

Ignoring the y-offset causes problems when the TwoLinesRenderer needs
to obey some other alignment, and will, therefore, be fixed in the
subsequent commits. To prevent those fixes from breaking the
MainIconView it is necessary to specify the desired vertical alignment.

https://bugzilla.gnome.org/show_bug.cgi?id=792665