meson: Add option to export symbols on MSVC-style builds
This adds an option that is enabled by default so that we can make Visual
Studio-style builds (which also includes clang-cl) export symbols from the
libgd copy that we incorporate into various things that we build or when
libgd is built standalone as a shared library.  Items such as gedit relies on
items in its included libgd parts to be exported in order to function
correctly.

This commit is the first part that adds a compiler macro which will activate
parts of the code so that symbols will be built with the needed symbols marked
for export, which will be in the subsequent commit(s).