libgd: Add header to define GD_API
This is the macro that will be used to decorate the symbols in the various
headers so that we can use compiler directives to export the symbols as
we incorporate libgd items in our GNOME items, as necessary.  This will
enable items such as gedit to work properly as they expect libgd items to
be exported as well during runtime.