libgd headers: Include gd-macros.h
...and mark the symbols (except the private ones) with GD_API.  This way, on
compilers that do not have an auto-export mechanism, we can export symbols by
defining GD_API as needed to use compiler directives to export the symbols.

This has been tested by generating the introspection files, as well as building
the two sample programs against a shared build of libgd with all features
enabled.