Pango 1.44 got the ability to automatically hyphenate on line breaks, which is enabled by default, but can be set off by a new attribute.
As a result, GdTwoLinesRenderer is now showing a hyphen in some cases
at end of first line, which is bad for filenames (like Totem or Evince
recent files view) because a filename may already include hyphens so
showing an extra hyphen which is not part of the filename is confusing.

This is a regression that Pango devs are not going to fix but instead
provided a new attribute to set this off.

So to fix this regression, let's restore previous behaviour when using
Pango >= 1.44

Fixes evince#1242
Fixes totem#349